# Frontend/RecRock-Calendar-Frontend/swagger.yaml
{
  "x-generator": "NSwag v13.15.9.0 (NJsonSchema v10.6.8.0 (Newtonsoft.Json v9.0.0.0))",
  "swagger": "2.0",
  "info": {
    "title": "RedRock Calendar Api",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/Events/{userReference}": {
      "get": {
        "tags": [
          "Events"
        ],
        "operationId": "Events_Get",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "userReference",
            "in": "path",
            "required": true,
            "format": "guid",
            "x-nullable": false
          },
          {
            "type": "string",
            "name": "date",
            "in": "query",
            "format": "date-time",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/EventDTO"
            }
          }
        }
      }
    },
    "/api/Events": {
      "post": {
        "tags": [
          "Events"
        ],
        "operationId": "Events_Post",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "name": "newEvent",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EventPostDTO"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/EventDTO"
            }
          }
        }
      },
      "get": {
        "tags": [
          "Events"
        ],
        "operationId": "Events_GetAll",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/EventDTO"
              }
            }
          }
        }
      }
    },
    "/api/Events/interval": {
      "get": {
        "tags": [
          "Events"
        ],
        "operationId": "Events_GetInterval",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "startDate",
            "in": "query",
            "format": "date-time",
            "x-nullable": false
          },
          {
            "type": "string",
            "name": "endDate",
            "in": "query",
            "format": "date-time",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/EventDTO"
              }
            }
          }
        }
      }
    },
    "/api/Events/{eventId}": {
      "delete": {
        "tags": [
          "Events"
        ],
        "operationId": "Events_Delete",
        "parameters": [
          {
            "type": "string",
            "name": "eventId",
            "in": "path",
            "required": true,
            "format": "guid",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
    "/api/Users": {
      "get": {
        "tags": [
          "Users"
        ],
        "operationId": "Users_Get",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/UserDTO"
              }
            }
          }
        }
      }
    },
    "/api/Users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "operationId": "Users_GetById",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true,
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/UserDTO"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "operationId": "Users_Update",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "guid",
            "x-nullable": false
          },
          {
            "name": "userDTO",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserUpdateDTO"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/UserDTO"
            }
          }
        }
      }
    },
    "/api/Users/{id}/changePassword": {
      "put": {
        "tags": [
          "Users"
        ],
        "operationId": "Users_ChangePassword",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "guid",
            "x-nullable": false
          },
          {
            "name": "passwords",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserChangePasswordDTO"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/UserDTO"
            }
          }
        }
      }
    },
    "/api/Users/login": {
      "post": {
        "tags": [
          "Users"
        ],
        "operationId": "Users_Login",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "name": "userParam",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserLoginDTO"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/UserDTO"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "EventDTO": {
      "type": "object",
      "required": [
        "id",
        "startDate",
        "endDate",
        "userReference"
      ],
      "properties": {
        "id": {
          "type": "string",
          "format": "guid"
        },
        "startDate": {
          "type": "string",
          "format": "date-time"
        },
        "endDate": {
          "type": "string",
          "format": "date-time"
        },
        "userReference": {
          "type": "string",
          "format": "guid"
        }
      }
    },
    "EventPostDTO": {
      "type": "object",
      "required": [
        "startDate",
        "endDate",
        "userReference"
      ],
      "properties": {
        "startDate": {
          "type": "string",
          "format": "date-time"
        },
        "endDate": {
          "type": "string",
          "format": "date-time"
        },
        "userReference": {
          "type": "string",
          "format": "guid"
        }
      }
    },
    "UserDTO": {
      "type": "object",
      "required": [
        "id",
        "fullName",
        "userName",
        "primaryColor",
        "secondaryColor"
      ],
      "properties": {
        "id": {
          "type": "string",
          "format": "guid",
          "minLength": 1
        },
        "fullName": {
          "type": "string",
          "minLength": 1
        },
        "userName": {
          "type": "string",
          "minLength": 1
        },
        "primaryColor": {
          "type": "string",
          "minLength": 1
        },
        "secondaryColor": {
          "type": "string",
          "minLength": 1
        }
      }
    },
    "UserChangePasswordDTO": {
      "type": "object",
      "properties": {
        "password": {
          "type": "string"
        },
        "newPassword": {
          "type": "string"
        },
        "newPasswordRepeat": {
          "type": "string"
        }
      }
    },
    "UserUpdateDTO": {
      "type": "object",
      "required": [
        "id"
      ],
      "properties": {
        "id": {
          "type": "string",
          "format": "guid"
        },
        "fullName": {
          "type": "string"
        },
        "primaryColor": {
          "type": "string"
        },
        "secondaryColor": {
          "type": "string"
        }
      }
    },
    "UserLoginDTO": {
      "type": "object",
      "required": [
        "userName",
        "password"
      ],
      "properties": {
        "userName": {
          "type": "string",
          "minLength": 1
        },
        "password": {
          "type": "string",
          "minLength": 1
        }
      }
    }
  }
}